Make array components distinct, refactor hooks
