+++ client/packages/programs/src/JsonForms/common/components/index.ts
@@ -5,7 +5,6 @@ export * from './Text';
 export * from './Select';
 export * from './Date';
 export * from './Array';
-export * from './FirstItemArray';
 export * from './Boolean';
 export * from './CategorizationTabLayout';
 export * from './Autocomplete';
