+++ client/packages/programs/src/JsonForms/common/index.ts
@@ -1,4 +1,4 @@
 export * from './JsonForm';
 export { JsonSchema, UISchemaElement } from '@jsonforms/core';
 export * from './styleConstants';
-export * from './useZodOptionsValidation';
+export * from './hooks/useZodOptionsValidation';
