+++ client/packages/programs/src/JsonForms/useJsonForms.tsx
@@ -31,8 +31,7 @@ import {
 import {
   DecisionTreeControl,
   decisionTreeTester,
-} from './components/DecisionTree';
-import { Note, noteTester } from './components/Note';
+} from './components/DecisionTree/DecisionTree';
 
 // https://stackoverflow.com/questions/57874879/how-to-treat-missing-undefined-properties-as-equivalent-in-lodashs-isequalwit
 // TODO: handle undefined and empty string as equal? e.g. initial data is undefined and current data is ""
@@ -93,7 +92,6 @@ const additionalRenderers: JsonFormsRendererRegistryEntry[] = [
     renderer: PreviousEncounterField,
   },
   { tester: decisionTreeTester, renderer: DecisionTreeControl },
-  { tester: noteTester, renderer: Note },
 ];
 
 export const useJsonForms = (
